The 'tardis2' protocol introduction. The 'tardis' datafile protocol is ambiguous as to the location to the file. There's the legacy system (ie. used by the synchrotorn) of /experimentid/filepath and the new one used by create_experiment and upload /experimentid/datasetid/filepath. The latter is recommended.
I plan on having the 'tardis' datafile protocol deprecated over time as no dataset folder in the store can cause conflicts and uncertainty.

To begin this I've created the 'tardis2' file protocol, used by create_experiment and upload functions. This specifies the /experimentid/datafileid/filepath structure in the store and will provide stability and certainty for the future.

In the future, a mets ingest mechanism should be created (Perhaps 1 web service for 'create experiment skeleton' that returns exp id and 1 for 'create dataset' that links to the experiment and returns a dataset id? That way we know where to send files for datasets..)